Fixed a bunch of gcc warnings from SDK
<==== Mercurial data ====>
Branch: play_sound_action

+++ common/src/plugins/camera_plugin_qt_wrapper.cpp
@@ -271,7 +271,7 @@ namespace nxcip_qt
         m_textBuf( new char[nxcip::MAX_TEXT_LEN] )
     {
         m_idsList = new char*[nxcip::MAX_RELAY_PORT_COUNT];
-        for( size_t i = 0; i < nxcip::MAX_RELAY_PORT_COUNT; ++i )
+        for( size_t i = 0; i < (uint)nxcip::MAX_RELAY_PORT_COUNT; ++i )
         {
             m_idsList[i] = new char[nxcip::MAX_ID_LEN];
             m_idsList[i][0] = 0;
@@ -280,7 +280,7 @@ namespace nxcip_qt
 
     CameraRelayIOManager::~CameraRelayIOManager()
     {
-        for( size_t i = 0; i < nxcip::MAX_RELAY_PORT_COUNT; ++i )
+        for( size_t i = 0; i < (uint)nxcip::MAX_RELAY_PORT_COUNT; ++i )
             delete[] m_idsList[i];
         delete[] m_idsList;
         m_idsList = NULL;
